# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for restaurant hall, orders, and warehouse management",
        "title": "Restaurant Management System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/audit": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audit"
                ],
                "summary": "List audit log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "table name",
                        "name": "table_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "record id",
                        "name": "record_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "from datetime (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "to datetime (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.AuditLog"
                            }
                        }
                    }
                }
            }
        },
...
        "/customers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "List customers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Customer"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Create customer",
                "parameters": [
                    {
                        "description": "customer",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Customer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Customer"
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Update customer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "customer id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "customer",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Customer"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "customers"
                ],
                "summary": "Delete customer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "customer id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/customers/{id}/total-spent": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get total spent by customer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "customer id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CustomerTotalSpent"
                        }
                    }
                }
            }
        },
        "/dishes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "List dishes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Dish"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Create or update dish",
                "parameters": [
                    {
                        "description": "dish",
                        "name": "dish",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Dish"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Dish"
                        }
                    }
                }
            }
        },
        "/dishes/available": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "List dishes available for ordering",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "max price",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DishAvailability"
                            }
                        }
                    }
                }
            }
        },
        "/dishes/{id}": {
            "delete": {
                "tags": [
                    "dishes"
                ],
                "summary": "Delete dish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "dish id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/dishes/{id}/ingredients": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Get dish ingredients",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "dish id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.DishWithIngredients"
                        }
                    }
                }
            }
        },
...
        "/import-errors": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import-errors"
                ],
                "summary": "List import errors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "entity",
                        "name": "entity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "from datetime (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "to datetime (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.ImportError"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AuditLog": {
            "type": "object",
            "properties": {
                "changed_at": {
                    "type": "string"
                },
                "changed_by": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "new_data": {
                    "type": "string"
                },
                "old_data": {
                    "type": "string"
                },
                "operation": {
                    "type": "string"
                },
                "record_id": {
                    "type": "integer"
                },
                "table_name": {
                    "type": "string"
                }
            }
        },
        "domain.Customer": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "vip_level": {
                    "type": "integer"
                }
            }
        },
        "domain.CustomerTotalSpent": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "total_spent": {
                    "type": "number"
                }
            }
        },
        "domain.Dish": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "cook_time_minutes": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "domain.DishAvailability": {
            "type": "object",
            "properties": {
                "all_products_available": {
                    "type": "boolean"
                },
                "can_be_ordered": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "domain.DishIngredientDetail": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "domain.DishWithIngredients": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "dish_id": {
                    "type": "integer"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.DishIngredientDetail"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "domain.ImportError": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "entity": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "raw_data": {
                    "type": "string"
                }
            }
        },
...
    }
}